+++ src/main/java/com/smeme/server/scheduler/DiaryScheduler.java
@@ -1,0 +1,19 @@
+package com.smeme.server.scheduler;
+
+import com.smeme.server.service.DiaryService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Component;
+
+@Component
+@EnableScheduling
+@RequiredArgsConstructor
+public class DiaryScheduler {
+    private final DiaryService diaryService;
+
+    @Scheduled(cron = "0 0 0 * * *")
+    public void deleteExpiredDiaries() {
+        diaryService.deleteExpiredDiary();
+    }
+}
